Make "get user from database" generic
